projet final (commentaire + renomme du solver)
